Flatten aesthetics into group

Occasionally, it may be necessary to enforce a group aesthetic from a selection of other aesthetics that are being passed to a packed layer. In this case, those aesthetics can be passed and preserved as an interaction term in the group aesthetic.

flatten_aesthetics_to_group(mapping, ...)

Arguments

mapping
a ggplot2 aesthetic mapping

...
aesthetic handles to flatten into the group aesthetic as interaction terms

Value

the mapping aesthetics with specified aesthetics stored additionally as interaction terms in the group aesthetics with any preexisting group aesthetics

Details

For example, if one would want to produce a line plot with an errorbar at every distinct point along the x-axis and use the linetype aesthetic to group the lines of the line plot, this will be problematic because it may also affect the linetype of the errorbars. To accommodate this, the linetype might be flattened into the group aesthetic, allowing for the linetype of the errorbars to be overridden without affecting the data groups.

Note

the x and y aesthetics will not be pulled into the group aesthetic

See also

allowed_aesthetics(), filter_aesthetics(), remove_aesthetics(), flatten_aesthetics_to_group(), add_eqv_aes(), split_aes_from_dots(), is_uneval()

Examples

library(ggplot2)
my.aes <- aes(x = mpg, y = wt, color = carb, fill = vs)
flatten_aesthetics_to_group(my.aes, 'color', 'fill')

#> * x      -> mpg
#> * y      -> wt
#> * colour -> carb
#> * fill   -> vs
#> * group  -> interaction(carb, vs)