Handling of argument passing to ggplots

A helper function to wrap ggplot calls, allowing for passing of a shared list of arguments prefixed by an id string. For example, a call to a custom ggplot template may allow for arguments 'line.color' and 'bar.color' to specify the line and bar colors separately. These arguments are parsed appropriately and passed to the appropriate sub-function.

ggpack(.call = NULL, ..., id = NULL, dots = NULL,
  auto_remove_aes = is.null(id) || any(sapply(id, is.null)),
  null_empty = FALSE, envir = parent.frame())

Arguments

.call
the ggplot geom function to invoke

...
additional arguments to use when calling the provided function

id
the prefix string to subset arguments by

dots
arguments to subset

auto_remove_aes
logical indicating whether aesthetic mappings that

null_empty
logical indicating whether NULL should be returned if no arguments are passed to call

envir
frame in which non-aesthetic arguments should be evaluated

Value

a call to the specified function with arguments subset for only those which match the specified prefix

Note

Ellipses arguments and the arguments passed through the dots args will overwrite previous arguments of the same name.

Examples

library(ggplot2)

# wrap layer calls in a ggpack to package them
ggpack(geom_bar(na.rm = TRUE))

#> ggpacket 
#> standalone plotting status: 
#>   Not all layers have been passed sufficient data 
#> 
#> [[1]]
#> Anonymous ggpacket layer
#> geom_bar(na.rm = TRUE) (LayerInstance):
#> 

# or more explicitly define layer properties
ggpack(geom_bar,              # layer function to bundle
  id = 'bar',                 # name of layer - used for filter listed args
  fill = 'red',               # ...'s before 'args' used as defaults
  dots = list(fill = 'blue'), # named 'args' param will overwrite prior ...'s
  fill = 'green')             # ...'s after 'args' will overwrite any others

#> ggpacket 
#> standalone plotting status: 
#>   Not all layers have been passed sufficient data 
#> 
#> [[1]]
#> ggpacket layer with id(s): bar 
#> geom_bar (function):
#>   call .. fill = "red"
#>   call .. fill = "green"
#>