Determine if an argument is an unevaluated expression

An argument is considered unevaluated if it returns TRUE for any of is.name(), is.call() or is.expression()

is_uneval(arg)

Arguments

arg
the input object to be tested

Value

TRUE if the value is of type name, call or expression

Note

This function may prove to be a heavy limitation of ggpackets if it is unable to accommodate custom unevaluated expression classes. Input on how best to handle such cases is welcome, but until a global solution is found, feedback on what other classes to consider is appreciated.

See also

Examples

ggpackets:::is_uneval(as.symbol('test'))

#> [1] TRUE

ggpackets:::is_uneval('test')

#> [1] FALSE