Remove specific aesthetics from mapping

Remove specified aesthetics by name from a ggplot2 aesthetic mapping, returning the same mapping with the specified aesthetics removed an preserved as an interaction term in the group aesthetic.

remove_aesthetics(mapping, ...)

Arguments

mapping
aesthetic mapping to use for filtering

...
mapping names to filter out

Value

an aesthetic mapping with filtered mappings removed and group mapping set as interaction terms of all non axial (non- x, y) terms.

See also

allowed_aesthetics(), filter_aesthetics(), remove_aesthetics(), flatten_aesthetics_to_group(), add_eqv_aes(), split_aes_from_dots(), is_uneval()

Examples

library(ggplot2)
remove_aesthetics(aes(x = wt, y = mpg, color = carb, fill = wt), 'fill')

#> * x      -> wt
#> * y      -> mpg
#> * colour -> carb
#> * group  -> interaction(wt)