Filter arguments into aesthetic and non-aesthetic

To accommodate passing of aesthetics as equivalent ellipses arguments without the need for wrapping in an aes call, it's often necessary to filter out any arguments which share the name of an acceptable aesthetic argument for the given Geom. In this case, arguments are split into two separate lists and returned within a list.

split_aes_from_dots(..., geom = NULL)

Arguments

...
arguments containing aesthetics mapping

geom
if specified, a ggplot2 Geom object can be passed to filter aesthetics for only a specified Geom.

Value

a list containing two named elements: 'aes' which contains the aesthetic mappings (of the Geom object if specified) as a pairlist with class uneval akin to the result of an aes call, and 'not_aes' which contains a pairlist of all other arguments

See also

allowed_aesthetics(), filter_aesthetics(), remove_aesthetics(), flatten_aesthetics_to_group(), add_eqv_aes(), split_aes_from_dots(), is_uneval()

Examples

library(ggplot2)
split_aes_from_dots(x = aes_x, my.arg.2 = 'my arg', color = aes_color,
  fill = 'aes_fill', my.arg.3 = list(1, 2, 3), geom = GeomBar)

#> $aes
#> * x      -> aes_x
#> * colour -> aes_color
#> * fill   -> "aes_fill"
#> 
#> $not_aes
#> * my.arg.2 -> "my arg"
#> * my.arg.3 -> list(1, 2, 3)
#>